Cleaned up the matrix solver code
