+++ examples/dc_test_inversion.py
@@ -270,140 +270,8 @@ print(f"Final model loss: {model_loss_history[-1]:.2e}")
 # =============================================================================
 
 # Get final inverted model
-inverted_model = forward_sim.inferred_model.detach().clone()
-
-# Create figure with multiple subplots
-fig, axes = plt.subplots(2, 3, figsize=(20, 12))
-
-# Plot 1: Loss history
-ax1 = axes[0, 0]
-iterations = range(len(loss_history))
-ax1.semilogy(iterations, loss_history, "b-", label="Total Loss")
-ax1.semilogy(iterations, data_loss_history, "r-", label="Data Loss")
-ax1.semilogy(iterations, model_loss_history, "g-", label="Model Loss")
-ax1.set_xlabel("Iteration")
-ax1.set_ylabel("Loss")
-ax1.set_title("Inversion Loss History")
-ax1.legend()
-ax1.grid(True)
-
-# Plot 2: True model cross-section
-ax2 = axes[0, 1]
-# Create 2D slice at y=0
-y_slice_idx = ny // 2
-true_model_2d = true_model.reshape(nx, ny, nz)[:, y_slice_idx, :]
-x_coords = mesh.nodes_x[:-1] + mesh.h[0] / 2
-z_coords = mesh.nodes_z[:-1] + mesh.h[2] / 2
-X, Z = torch.meshgrid(x_coords, z_coords, indexing="ij")
-
-im2 = ax2.pcolormesh(
-    X.cpu(),
-    Z.cpu(),
-    true_model_2d.cpu(),
-    norm=LogNorm(vmin=10, vmax=1000),
-    cmap="viridis",
-)
-ax2.set_xlabel("X (m)")
-ax2.set_ylabel("Z (m)")
-ax2.set_title("True Model (y=0 slice)")
-ax2.set_aspect("equal")
-plt.colorbar(im2, ax=ax2, label="Resistivity (Ω⋅m)")
-
-# Plot 3: Inverted model cross-section
-ax3 = axes[1, 0]
-inverted_full = active_mapping.forward(inverted_model)
-inverted_model_2d = inverted_full.reshape(nx, ny, nz)[:, y_slice_idx, :]
-
-im3 = ax3.pcolormesh(
-    X.cpu(),
-    Z.cpu(),
-    inverted_model_2d.cpu(),
-    norm=LogNorm(vmin=10, vmax=1000),
-    cmap="viridis",
-)
-ax3.set_xlabel("X (m)")
-ax3.set_ylabel("Z (m)")
-ax3.set_title("Inverted Model (y=0 slice)")
-ax3.set_aspect("equal")
-plt.colorbar(im3, ax=ax3, label="Resistivity (Ω⋅m)")
-
-# Plot 4: Anomaly profile through Y-slices
-ax4 = axes[1, 1]
-# Create anomaly profile showing most anomalous value for each Y slice
-background_rho = 100.0  # Background resistivity
-
-# For true model
-true_full = true_model.reshape(nx, ny, nz)
-anomaly_profile_true = []
-for y_idx in range(ny):
-    y_slice = true_full[:, y_idx, :]
-    # Find most anomalous value (furthest from background)
-    deviation = torch.abs(y_slice - background_rho)
-    max_deviation_idx = torch.argmax(deviation)
-    most_anomalous_value = y_slice.flatten()[max_deviation_idx]
-    anomaly_profile_true.append(most_anomalous_value.item())
-
-# For inverted model
-inverted_full = active_mapping.forward(inverted_model)
-inverted_full_reshaped = inverted_full.reshape(nx, ny, nz)
-anomaly_profile_inv = []
-for y_idx in range(ny):
-    y_slice = inverted_full_reshaped[:, y_idx, :]
-    # Find most anomalous value (furthest from background)
-    deviation = torch.abs(y_slice - background_rho)
-    max_deviation_idx = torch.argmax(deviation)
-    most_anomalous_value = y_slice.flatten()[max_deviation_idx]
-    anomaly_profile_inv.append(most_anomalous_value.item())
-
-# Y coordinates for plotting
-y_centers = mesh.cell_centers_y.cpu().numpy()
-
-# Plot anomaly profiles
-ax4.plot(y_centers, anomaly_profile_true, "b-", linewidth=2, label="True Model")
-ax4.plot(y_centers, anomaly_profile_inv, "r--", linewidth=2, label="Inverted Model")
-ax4.axhline(y=background_rho, color="k", linestyle=":", alpha=0.5, label="Background")
-ax4.set_xlabel("Y (m)")
-ax4.set_ylabel("Most Anomalous Resistivity (Ω⋅m)")
-ax4.set_title("Anomaly Profile Through Y-Slices")
-ax4.legend()
-ax4.grid(True, alpha=0.3)
-
-# Plot 5: Data fit comparison
-ax5 = axes[1, 2]
-final_dpred = forward_sim.simulation.dpred(active_mapping.forward(inverted_model))
-ax5.loglog(
-    torch.abs(dpred_reference).cpu(),
-    torch.abs(final_dpred.detach()).cpu(),
-    "bo",
-    alpha=0.6,
-)
-data_min = min(torch.abs(dpred_reference).min(), torch.abs(final_dpred.detach()).min())
-data_max = max(torch.abs(dpred_reference).max(), torch.abs(final_dpred.detach()).max())
-ax5.loglog([data_min, data_max], [data_min, data_max], "r--", label="Perfect fit")
-ax5.set_xlabel("True Data (V)")
-ax5.set_ylabel("Predicted Data (V)")
-ax5.set_title("Data Fit")
-ax5.legend()
-ax5.grid(True)
-
-# Plot 6: Model comparison scatter plot
-ax6 = axes[0, 2]
-ax6.scatter(
-    active_model.detach().cpu().numpy(),
-    inverted_model.detach().cpu().numpy(),
-    alpha=0.5,
-)
-ax6.plot([10, 1000], [10, 1000], "r--", label="Perfect fit")
-ax6.set_xlabel("True Resistivity (Ω⋅m)")
-ax6.set_ylabel("Inverted Resistivity (Ω⋅m)")
-ax6.set_title("Model Comparison")
-ax6.legend()
-ax6.grid(True)
-ax6.set_xscale("log")
-ax6.set_yscale("log")
-
-plt.tight_layout()
-plt.show()
+inverted_model = forward_sim.inferred_model.detach().clone().cpu()
+true_model = true_model.detach().cpu()
 
 # =============================================================================
 # Print summary statistics
