+++ simpegtorch/torchmatsolver/__init__.py
@@ -3,7 +3,7 @@
 # Import necessary modules or classes from this package
 # Example:
 # from .module_name import ClassName, function_name
-from .torchmatsolver import TorchMatSolver
-from .torchMUMPSsolver import TorchMUMPSsolver
+from .torchmatsolver import TorchMatSolver, batched_sparse_solve
+from .torchMUMPSsolver import TorchMUMPSsolver, batched_mumps_solve
 
 # You can also define package-level variables or functions here if needed
